rework and fix the synchronization between joined joycons and fix some bugs
When joycons are joined, the left joycon acts as the main controller for storing inputs and sending them to its virtual controller. An update of the inputs of the right joycon will instantly trigger an update of the virtual controller inputs. IMU datas are still sent independantly. Actions related to inputs/IMU that must be done only once will happens eitehr on the main controller or on the controller set as the main one for the IMU.